mcpserver/agent_bilibili_video/requirements.txt
bilibili-api-python>=17.3.0